DigitalShadows(PAPP-11277): Merged the customer code and the old code present in phantomCyber repo

+++ Apps/phdigitalshadows/dsapi/model/data_breach.py
@@ -2,7 +2,7 @@
 # Copyright (c) 2017 Digital Shadows Ltd.
 #
 
-from .ds_model import DSModel
+from ds_model import DSModel
 
 
 class DataBreach(DSModel):
@@ -97,18 +97,18 @@ class DataBreach(DSModel):
     @classmethod
     def from_json(cls, json):
         cast = DSModel.cast
-        return cls(cast(json.get('id'), int),
+        return cls(cast(json.get('id'), long),
                    json.get('title'),
                    json.get('domainName'),
                    json.get('occurred'),
                    json.get('modified'),
                    json.get('published'),
-                   cast(json.get('incident').get('id'), int),
+                   cast(json.get('incident').get('id'), long),
                    json.get('incident').get('scope'),
                    json.get('incident').get('type'),
                    json.get('incident').get('severity'),
                    json.get('incident').get('title'),
-                   cast(json.get('domainCount'), int),
-                   cast(json.get('recordCount'), int),
+                   cast(json.get('domainCount'), long),
+                   cast(json.get('recordCount'), long),
                    json.get('sourceUrl'),
-                   cast(json.get('organisationUsernameCount'), int), json)
+                   cast(json.get('organisationUsernameCount'), long), json)
