+++ Apps/phdigitalshadows/dsapi/model/data_breach_record.py
@@ -2,7 +2,7 @@
 # Copyright (c) 2017 Digital Shadows Ltd.
 #
 
-from .ds_model import DSModel
+from ds_model import DSModel
 
 
 class DataBreachRecord(DSModel):
@@ -61,7 +61,7 @@ class DataBreachRecord(DSModel):
     @classmethod
     def from_json(cls, json):
         cast = DSModel.cast
-        return cls(cast(json.get('id'), int),
+        return cls(cast(json.get('id'), long),
                    json.get('username'),
                    json.get('password'),
                    json.get('review'),
