+++ Apps/phdigitalshadows/dsapi/model/incident.py
@@ -44,5 +44,5 @@ class Incident(DSModel):
 
     @classmethod
     def from_json(cls, json):
-        incident_id = DSModel.cast(json.get('id'), int)
+        incident_id = DSModel.cast(json.get('id'), long)
         return cls(incident_id, json)
