+++ Apps/phdigitalshadows/dsapi/model/intelligence_incident.py
@@ -44,6 +44,6 @@ class IntelligenceIncident(DSModel):
 
     @classmethod
     def from_json(cls, json):
-        incident_id = DSModel.cast(json.get('id'), int)
+        incident_id = DSModel.cast(json.get('id'), long)
         return cls(incident_id, json)
